% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomoments.R
\name{ctmc.sim}
\alias{ctmc.sim}
\title{Simulation of a CTMC path.}
\usage{
ctmc.sim(t, rate.mat, states = c("a","c","g","t"), init.state)
}
\arguments{
\item{t}{A nonnegative numeric scalar representing the CTMC time interval 
length.}

\item{rate.mat}{A CTMC rate matrix.}

\item{states}{A character or integer vector denoting the CTMC state space. 
Defaults to the set of DNA nucleotides \code{c("a","c","g","t")}.}

\item{init.state}{The initial state of the CTMC.  This state must be an 
element of \code{states}.}
}
\value{
A data frame with two named columns \code{"state"} and \code{"time"}.
  The \code{"state"} column stores the observed states of the CTMC path and 
  the \code{"time"} column specifies the times at which each of the observed 
  states is entered.
}
\description{
Generates a realization of a continuous-time Markov chain (CTMC).
}
\seealso{
\code{\link{moments.ctmcjumps}}
}

