% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomoments.R
\name{int.states.sim}
\alias{int.states.sim}
\title{Simulation of internal node states on a phylogenetic tree.}
\usage{
int.states.sim(tree, rate.mat, root.dist, scale = FALSE,
               states = c("a","c","g","t"), tip.data, N)
}
\arguments{
\item{tree}{A tree object of class \code{"phylo"} (\strong{ape} format).}

\item{rate.mat}{A continuous-time Markov chain (CTMC) rate matrix.}

\item{root.dist}{A numeric vector defining the root distribution of the 
evolutionary process on the tree. If this process starts at stationarity 
(as is commonly assumed), then the root distribution is equal to the CTMC 
stationary distribution.}

\item{scale}{A logical indicating whether to scale the time dimension.  If 
\code{TRUE}, then time is specified in terms of the expected number of CTMC
substitutions per site.}

\item{states}{A character or integer vector denoting the CTMC state space. 
Defaults to the set of DNA nucleotides \code{c("a","c","g","t")}.}

\item{tip.data}{A character or integer vector of observed tip states.  These 
tip states must be contained in the vector \code{states}.}

\item{N}{An integer specifying the number of samples to draw.}
}
\value{
A character or integer matrix with dimensions 
  \code{c(2*length(tree$tip.label)-1, N)}, where each column contains an 
  independent sample of internal node states along with the observed tip 
  states.
}
\description{
Samples internal node states on a phylogeny conditional on the observed tip 
states.
}
\details{
This simulation function follows the procedure outlined in Nielsen (2002).

More information on phylogenetic tree objects of class \code{"phylo"} can be 
found at \url{http://ape-package.ird.fr/misc/FormatTreeR_24Oct2012.pdf}.
}
\references{
Nielsen R (2002) \dQuote{Mapping mutations on phylogenies,} 
  \emph{Systematic Biology,} 51(5):729-739.
}
\seealso{
\code{\link{tips.sim}}, \code{\link{phylojumps.sim}}
}

