% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomoments.R
\name{joint.postmean.moments.phylojumps}
\alias{joint.postmean.moments.phylojumps}
\title{Joint moments of the posterior mean of labeled substitution counts on a 
phylogenetic tree.}
\usage{
joint.postmean.moments.phylojumps(tree, rate.mat, label.mat,
                                  edge.set1 = NULL, edge.set2 = NULL, 
                                  subtree = NULL, root.dist,
                                  scale = FALSE, N = 100000)
}
\arguments{
\item{tree}{A tree object of class \code{"phylo"} (\strong{ape} format).}

\item{rate.mat}{A reversible continuous-time Markov chain (CTMC) rate matrix.}

\item{label.mat}{A 0-1 matrix of the same size as \code{rate.mat}.  This 
matrix labels the substitutions of interest.}

\item{edge.set1, edge.set2}{Two integer vectors of branch indices specifying 
the branch sets over which the joint posterior moments will be calculated.}

\item{subtree}{An integer denoting a node index in the tree.  If provided, 
then \code{edge.set1} will consist of the branches in the subtree beneath 
the specified node (including the branch above \code{subtree}), while 
\code{edge.set2} will contain the branches in the complementary supertree.}

\item{root.dist}{A numeric vector defining the root distribution of the 
evolutionary process on the tree. If this process starts at stationarity 
(as is commonly assumed), then the root distribution is equal to the CTMC 
stationary distribution.}

\item{scale}{A logical indicating whether to scale the time dimension.  If 
\code{TRUE}, then time is specified in terms of the expected number of CTMC
substitutions per site.}

\item{N}{An integer specifying the number of Monte Carlo samples to draw.}
}
\value{
A numeric vector holding the means, variances, and covariance of the 
  posterior mean of labeled substitution counts.
}
\description{
This function calculates the means, variances, and covariance of the 
posterior mean of labeled substitution counts over two prespecified sets of 
branches on a phylogeny using a simulation-free dynamic programming algorithm
and Monte Carlo sampling.
}
\details{
The means obtained in this function are computed exactly, while the variances
and covariance are partially approximated using Monte Carlo simulations.

More information on phylogenetic tree objects of class \code{"phylo"} can be 
found at \url{http://ape-package.ird.fr/misc/FormatTreeR_24Oct2012.pdf}.
}
\references{
Dhar A and Minin VN (2017) \dQuote{Calculating Higher-Order 
  Moments of Phylogenetic Stochastic Mapping Summaries in Linear Time}, 
  \emph{Journal of Computational Biology}, 24(5):377-399.
}
\seealso{
\code{\link{joint.prior.moments.phylojumps}}, 
  \code{\link{joint.post.moments.phylojumps}}, 
  \code{\link{postmean.moments.phylojumps}}
}

