% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomoments.R
\name{moments.ctmcjumps}
\alias{moments.ctmcjumps}
\title{Restricted factorial moments of labeled substitution counts for a reversible 
CTMC.}
\usage{
moments.ctmcjumps(t, rate.mat, label.mat,
                  root.dist = NULL, scale = FALSE)
}
\arguments{
\item{t}{A nonnegative numeric vector of CTMC time interval lengths.}

\item{rate.mat}{A reversible CTMC rate matrix.}

\item{label.mat}{A 0-1 matrix of the same size as \code{rate.mat}.  This 
matrix labels the substitutions of interest.}

\item{root.dist}{A numeric vector defining the CTMC stationary distribution. 
This only needs to be specified when \code{scale = TRUE}.}

\item{scale}{A logical indicating whether to scale the time dimension.  If 
\code{TRUE}, then time is specified in terms of the expected number of CTMC
substitutions.}
}
\value{
A list with named elements \code{"zeroth"}, \code{"first"}, and 
  \code{"second"}.  Each list element is a three-dimensional array with 
  dimensions \code{c(nrow(rate.mat), ncol(rate.mat), length(t))}.  The 
  \code{"zeroth"}, \code{"first"}, and \code{"second"} elements of this list 
  contain the 0th, 1st, and 2nd restricted factorial moment matrices of 
  labeled CTMC substitution counts, respectively, for the time interval 
  lengths defined by \code{t}.
}
\description{
This function calculates the 0th, 1st, and 2nd restricted factorial moment 
matrices of labeled substitution counts for a reversible continuous-time 
Markov chain (CTMC) model.
}
\details{
The 0th restricted factorial moment matrix of labeled CTMC substitution 
counts is defined as the CTMC transition probability matrix.
}
\references{
V.N. Minin and M.A. Suchard (2008) \dQuote{Counting labeled 
  transitions in continuous-time Markov models of evolution,} \emph{Journal 
  of Mathematical Biology,} 56(3):391-412.
}
\seealso{
\code{\link{ctmc.sim}}
}

