% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomoments.R
\name{phylojumps.sim}
\alias{phylojumps.sim}
\title{Simulation of substitution counts on a phylogenetic tree.}
\usage{
phylojumps.sim(tree, rate.mat, root.dist, scale = FALSE,
               states = c("a","c","g","t"), seq.data, N)
}
\arguments{
\item{tree}{A tree object of class \code{"phylo"} (\strong{ape} format).}

\item{rate.mat}{A continuous-time Markov chain (CTMC) rate matrix.}

\item{root.dist}{A numeric vector defining the root distribution of the 
evolutionary process on the tree. If this process starts at stationarity 
(as is commonly assumed), then the root distribution is equal to the CTMC 
stationary distribution.}

\item{scale}{A logical indicating whether to scale the time dimension.  If 
\code{TRUE}, then time is specified in terms of the expected number of CTMC
substitutions per site.}

\item{states}{A character or integer vector denoting the CTMC state space. 
Defaults to the set of DNA nucleotides \code{c("a","c","g","t")}.}

\item{seq.data}{A character or integer matrix representing the observed 
sequence alignment with rows (columns) that correspond to the sequences 
(sites) under study.  All elements of \code{seq.data} must be contained in 
the vector \code{states}.}

\item{N}{An integer specifying the number of samples to draw.}
}
\value{
An integer matrix with dimensions \code{c(N, ncol(seq.data))}, where 
  the \emph{i}'th column contains \code{N} independent samples of 
  substitution counts for the \emph{i}'th site in \code{seq.data}.
}
\description{
Simulates the number of substitutions on a phylogeny conditional on the 
observed tip states.
}
\details{
This simulation function follows the procedure outlined in Nielsen (2002).

More information on phylogenetic tree objects of class \code{"phylo"} can be 
found at \url{http://ape-package.ird.fr/misc/FormatTreeR_24Oct2012.pdf}.
}
\references{
Nielsen R (2002) \dQuote{Mapping mutations on phylogenies}, 
  \emph{Systematic Biology}, 51(5):729-739.
}
\seealso{
\code{\link{int.states.sim}}, \code{\link{tips.sim}}
}

