% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomoments.R
\docType{package}
\name{phylomoments}
\alias{phylomoments}
\alias{phylomoments-package}
\title{phylomoments: Efficient computations of phylogenetic stochastic mapping 
moments.}
\description{
This package implements a simulation-free dynamic programming algorithm for 
calculating higher-order moments of stochastic mapping summaries on a 
phylogeny. Utility functions related to simulation-based stochastic mapping 
and first-order moments are also provided.
}
\section{}{
 Please see the vignettes for some examples of using the functions 
  provided in this package.  They can be accessed using 
  \code{browseVignettes("phylomoments")}.
}
\author{
Amrit Dhar, Vladimir N. Minin
  
  Maintainer: Amrit Dhar \email{adhar@uw.edu}
}
\references{
Nielsen R (2002) \dQuote{Mapping mutations on phylogenies,} 
  \emph{Systematic Biology,} 51(5):729-739.
  
  Minin VN and Suchard MA (2008) \dQuote{Counting labeled transitions in 
  continuous-time Markov models of evolution,} \emph{Journal of Mathematical 
  Biology,} 56(3):391-412.
  
  OUR PAPER!!!
}

