% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomoments.R
\name{postmean.moments.phylojumps}
\alias{postmean.moments.phylojumps}
\title{Moments of the posterior mean of labeled substitution counts on a 
phylogenetic tree.}
\usage{
postmean.moments.phylojumps(tree, rate.mat, label.mat,
                            edge.set, root.dist,
                            scale = FALSE, N = 100000)
}
\arguments{
\item{tree}{A tree object of class \code{"phylo"} (\strong{ape} format).}

\item{rate.mat}{A reversible continuous-time Markov chain (CTMC) rate matrix.}

\item{label.mat}{A 0-1 matrix of the same size as \code{rate.mat}.  This 
matrix labels the substitutions of interest.}

\item{edge.set}{An integer vector of branch indices specifying the branch set
over which the moments will be calculated.}

\item{root.dist}{A numeric vector defining the root distribution of the 
evolutionary process on the tree. If this process starts at stationarity 
(as is commonly assumed), then the root distribution is equal to the CTMC 
stationary distribution.}

\item{scale}{A logical indicating whether to scale the time dimension.  If 
\code{TRUE}, then time is specified in terms of the expected number of CTMC
substitutions per site.}

\item{N}{An integer specifying the number of Monte Carlo samples to draw.}
}
\value{
A numeric vector holding the mean and variance of the posterior mean 
  of labeled substitution counts.
}
\description{
This function calculates the mean and variance of the posterior mean of 
labeled substitution counts over a prespecified set of branches on a 
phylogeny using a simulation-free dynamic programming algorithm and Monte 
Carlo sampling.
}
\details{
The mean obtained in this function is computed exactly, while the variance is
partially approximated using Monte Carlo simulations.

More information on phylogenetic tree objects of class \code{"phylo"} can be 
found at \url{http://ape-package.ird.fr/misc/FormatTreeR_24Oct2012.pdf}.
}
\references{
Dhar A and Minin VN (2017) \dQuote{Calculating Higher-Order 
  Moments of Phylogenetic Stochastic Mapping Summaries in Linear Time}, 
  \emph{Journal of Computational Biology}, 24(5):377-399.
}
\seealso{
\code{\link{prior.moments.phylojumps}}, 
  \code{\link{post.moments.phylojumps}}, 
  \code{\link{joint.postmean.moments.phylojumps}}
}

