% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomoments.R
\name{prior.moments.phylojumps}
\alias{prior.moments.phylojumps}
\title{Prior moments of labeled substitution counts on a phylogenetic tree.}
\usage{
prior.moments.phylojumps(tree, rate.mat, label.mat,
                         edge.set, root.dist, scale = FALSE)
}
\arguments{
\item{tree}{A tree object of class \code{"phylo"} (\strong{ape} format).}

\item{rate.mat}{A reversible continuous-time Markov chain (CTMC) rate matrix.}

\item{label.mat}{A 0-1 matrix of the same size as \code{rate.mat}.  This 
matrix labels the substitutions of interest.}

\item{edge.set}{An integer vector of branch indices specifying the branch set
over which the prior moments will be calculated.}

\item{root.dist}{A numeric vector defining the root distribution of the 
evolutionary process on the tree. If this process starts at stationarity 
(as is commonly assumed), then the root distribution is equal to the CTMC 
stationary distribution.}

\item{scale}{A logical indicating whether to scale the time dimension.  If 
\code{TRUE}, then time is specified in terms of the expected number of CTMC
substitutions per site.}
}
\value{
A numeric vector holding the prior mean and variance of labeled 
  substitution counts.
}
\description{
This function computes the prior mean and variance of labeled substitution 
counts over a prespecified set of branches on a phylogeny using a 
simulation-free dynamic programming algorithm.
}
\details{
More information on phylogenetic tree objects of class \code{"phylo"} can be 
found at \url{http://ape-package.ird.fr/misc/FormatTreeR_24Oct2012.pdf}.
}
\references{
OUR PAPER!!!
}
\seealso{
\code{\link{post.moments.phylojumps}}, 
  \code{\link{postmean.moments.phylojumps}}, 
  \code{\link{joint.prior.moments.phylojumps}}
}

