% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomoments.R
\name{tips.sim}
\alias{tips.sim}
\title{Simulation of terminal node states on a phylogenetic tree.}
\usage{
tips.sim(tree, rate.mat, root.dist, scale = FALSE,
         states = c("a","c","g","t"), N)
}
\arguments{
\item{tree}{A tree object of class \code{"phylo"} (\strong{ape} format).}

\item{rate.mat}{A continuous-time Markov chain (CTMC) rate matrix.}

\item{root.dist}{A numeric vector defining the root distribution of the 
evolutionary process on the tree. If this process starts at stationarity 
(as is commonly assumed), then the root distribution is equal to the CTMC 
stationary distribution.}

\item{scale}{A logical indicating whether to scale the time dimension.  If 
\code{TRUE}, then time is specified in terms of the expected number of CTMC
substitutions per site.}

\item{states}{A character or integer vector denoting the CTMC state space. 
Defaults to the set of DNA nucleotides \code{c("a","c","g","t")}.}

\item{N}{An integer specifying the number of samples to draw.}
}
\value{
A character or integer matrix with dimensions 
  \code{c(length(tree$tip.label), N)}, where each column contains an 
  independent sample of terminal node states.
}
\description{
Simulates tip states on a phylogeny according to a continuous-time Markov 
model of evolution.
}
\details{
More information on phylogenetic tree objects of class \code{"phylo"} can be 
found at \url{http://ape-package.ird.fr/misc/FormatTreeR_24Oct2012.pdf}.
}
\seealso{
\code{\link{int.states.sim}}, \code{\link{phylojumps.sim}}
}

