% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylomoments.R
\docType{package}
\name{phylomoments}
\alias{phylomoments}
\alias{phylomoments-package}
\title{phylomoments: Efficient computations of phylogenetic stochastic mapping 
moments.}
\description{
This package implements a simulation-free dynamic programming algorithm for 
calculating higher-order moments of stochastic mapping summaries on a 
phylogeny. Utility functions related to simulation-based stochastic mapping 
and continuous-time Markov chains (CTMCs) are also provided.
}
\section{}{
 Please see the vignettes for some examples of using the functions 
  provided in this package.
}
\author{
Amrit Dhar \email{adhar@uw.edu}
}
\references{
OUR PAPER!!!
  
  R. Nielsen (2002) \dQuote{Mapping mutations on phylogenies,} 
  \emph{Systematic Biology,} 51(5):729-739.
  
  V.N. Minin and M.A. Suchard (2008) \dQuote{Counting labeled transitions in 
  continuous-time Markov models of evolution,} \emph{Journal of Mathematical 
  Biology,} 56(3):391-412.
}

